Update generated SDKs

+++ lib/commercetools-api/src/Models/Order/LineItemImportDraftBuilder.php
@@ -53,7 +53,7 @@ final class LineItemImportDraftBuilder implements Builder
     private $price;
 
     /**
-     * @var ?float
+     * @var ?int
      */
     private $quantity;
 
@@ -125,7 +125,7 @@ final class LineItemImportDraftBuilder implements Builder
     }
 
     /**
-     * @return null|float
+     * @return null|int
      */
     public function getQuantity()
     {
@@ -236,10 +236,10 @@ final class LineItemImportDraftBuilder implements Builder
     }
 
     /**
-     * @param ?float $quantity
+     * @param ?int $quantity
      * @return $this
      */
-    public function withQuantity(?float $quantity)
+    public function withQuantity(?int $quantity)
     {
         $this->quantity = $quantity;
